Make the plugin work as a Text Formatting option
When using a plugin as a custom field "Text Formatting" option, EE passes the field's content to the plugin as a parmeter. 
This adds that parameter and uses it if it exists, allowing a custom field to automatically run its content through the plugin without the need for a the template tag pair.